backend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.9
inflection==0.5.1